render: Move render-specific stuff to pixel_bender_support.rs
This way pixel_bender.rs is self contained.
